+++ modules/segmentation_module/segmentation.py
@@ -88,7 +88,7 @@ class Segmentation(QWidget):
         Retrieve the input parameters
         Iterate over the image paths given performing f.main() function
         """
-        def check_inputs(image_paths, model_path):
+        def check_inputs(image_paths, model_path, output_paths, output_basenames):
             if len(image_paths) == 0:
                 self.logger.error('Image missing')
                 return False
@@ -108,26 +108,32 @@ class Segmentation(QWidget):
                 self.logger.error('Output folder missing')
                 self.output_folder.setFocus()
                 return False
+            output_files = [os.path.join(d, f) for d, f in zip(output_paths, output_basenames)]
+            duplicates = [x for x, y in zip(image_paths, output_files) if output_files.count(y) > 1]
+            if len(duplicates) > 0:
+                self.logger.error('More than one input file will output to the same file.\nTo avoid overwriting output files, either use input image folder as output folder or do not process images from different input directories with same name.\nProblematic input files:\n%s', '\n'.join(duplicates[:4] + (['...'] if len(duplicates) > 4 else [])))
+                return False
             return True
 
         image_paths = self.image_list.get_file_list()
         model_path = self.selected_model.text()
+        output_basenames = [os.path.splitext(os.path.basename(path))[0] for path in image_paths]
+        if self.use_input_folder.isChecked():
+            output_paths = [os.path.join(os.path.dirname(path), 'segmentation_masks') for path in image_paths]
+        else:
+            output_paths = [self.output_folder.text() for path in image_paths]
 
-        if not check_inputs(image_paths, model_path):
+        if not check_inputs(image_paths, model_path, output_paths, output_basenames):
             return
 
-        for image_path in image_paths:
+        for image_path, output_path, output_basename in zip(image_paths, output_paths, output_basenames):
             if os.path.isfile(image_path):
-                if self.use_input_folder.isChecked():
-                    output_path = os.path.join(os.path.dirname(image_path), 'segmentation_masks')
-                else:
-                    output_path = self.output_folder.text()
                 self.logger.info("Segmenting image %s", image_path)
                 QApplication.setOverrideCursor(QCursor(Qt.BusyCursor))
                 QApplication.processEvents()
 
                 try:
-                    f.main(image_path, model_path, output_path, self.display_results.isChecked())
+                    f.main(image_path, model_path, output_path, output_basename, self.display_results.isChecked())
                 except Exception as e:
                     self.logger.error("Segmentation failed.\n%s", str(e))