+++ modules/segmentation_module/segmentation_functions.py
@@ -6,11 +6,14 @@ import tifffile
 from cellpose import models
 from cellpose import version_str as cellpose_version
 from general import general_functions as gf
+from PyQt5.QtGui import QCursor
+from PyQt5.QtCore import Qt
+from PyQt5.QtWidgets import QMessageBox
 
 
-def main(image_path, model_path, output_path, display_results=True, use_gpu=True):
+def main(image_path, model_path, output_path, output_basename, display_results=True, use_gpu=True):
     """
-    Load image, segment with cellpose and save the resulting mask 
+    Load image, segment with cellpose and save the resulting mask
     into `output_path` directory using filename <image basename>_mask.tif
     Note : we assume that the image first channel is ALWAYS BF and we will only apply the segmentation on that channel
 
@@ -22,6 +25,8 @@ def main(image_path, model_path, output_path, display_results=True, use_gpu=True
         cellpose pretrained model path
     output_path: str
         output directory
+    output_basename: str
+        output basename. Output file will be saved as `output_path`/`output_basename`_mask.tif and `output_path`/`output_basename`.log.
     display_results: bool, default True
         display input image and segmentation mask in napari
     use_gpu: bool, default False
@@ -36,7 +41,7 @@ def main(image_path, model_path, output_path, display_results=True, use_gpu=True
         os.makedirs(output_path)
 
     # Log to file
-    logfile = os.path.join(output_path, os.path.splitext(os.path.basename(image_path))[0]+".log")
+    logfile = os.path.join(output_path, output_basename+".log")
     logger.setLevel(logging.DEBUG)
     logger.debug("writing log output to: %s", logfile)
     logfile_handler = logging.FileHandler(logfile, mode='w')
@@ -44,7 +49,7 @@ def main(image_path, model_path, output_path, display_results=True, use_gpu=True
     logfile_handler.setLevel(logging.INFO)
     logger.addHandler(logfile_handler)
 
-    ## Cellpose_version already contains platform, python version and torch version
+    # Cellpose_version already contains platform, python version and torch version
     logger.info("System info: %s\nnumpy version: %s\nnapari version: %s", cellpose_version, np.__version__, napari.__version__)
     logger.info("image: %s", image_path)
     logger.info("cellpose model: %s", model_path)
@@ -58,28 +63,24 @@ def main(image_path, model_path, output_path, display_results=True, use_gpu=True
         image = gf.Image(image_path)
         image.imread()
     except Exception as e:
-        logging.getLogger(__name__).error('Error loading image '+image_path+'\n'+str(e))
+        logging.getLogger(__name__).error('Error loading image %s\n%s', image_path, str(e))
 
     # Create cellpose model
     logger.debug("loading cellpose model %s", model_path)
     model = models.CellposeModel(gpu=use_gpu, pretrained_model=model_path)
-    
+
     tot_iterations = image.sizes['Z']*image.sizes['T']*image.sizes['F']
 
-    if display_results: # Open image in napari
-        # Only import PyQt5 if needed (no need for PyQt5 dependency if display_results is False)
-        from PyQt5.QtGui import QCursor
-        from PyQt5.QtCore import Qt
-        from PyQt5.QtWidgets import QMessageBox
-
+    if display_results:
+        # Open image in napari
         viewer_images = napari.Viewer(title=image_path)
         image_napari = image.get_TYXarray()
-        image_napari = image_napari[:, np.newaxis, : ,:]
+        image_napari = image_napari[:, np.newaxis, :, :]
         viewer_images.add_image(image_napari, name="Input image")
-        
-        ## TODO: find a way to use logging package instead?
+
+        # TODO: find a way to use logging package instead?
         # Setup logging into napari window.
-        
+
         # Set cursor to BusyCursor
         napari.qt.get_app().setOverrideCursor(QCursor(Qt.BusyCursor))
         napari.qt.get_app().processEvents()
@@ -89,7 +90,7 @@ def main(image_path, model_path, output_path, display_results=True, use_gpu=True
 
     # Cellpose segmentation
     logger.info("Cellpose segmentation (model diameter=%s)", model.diam_labels)
-    
+
     iteration = 0
     multiple_fov = True if image.sizes['F'] > 1 else False
 
@@ -98,24 +99,25 @@ def main(image_path, model_path, output_path, display_results=True, use_gpu=True
         for t in range(image.sizes['T']):
             # Always assuming BF in channel=0 and only one Z channel
             iteration += 1
-            image_2D = image.image[f, t, 0, 0, : ,:]
-            if display_results: # Logging into napari window
+            image_2D = image.image[f, t, 0, 0, :, :]
+            if display_results:
+                # Logging into napari window
                 pbr.set_description(f"cellpose segmentation {iteration}/{tot_iterations}")
                 pbr.update(1)
             logger.info("cellpose segmentation %s/%s", iteration, tot_iterations)
-            mask[t,:,:], _, _ = model.eval(image_2D, diameter=model.diam_labels, channels=[0, 0])
+            mask[t, :, :], _, _ = model.eval(image_2D, diameter=model.diam_labels, channels=[0, 0])
 
         # Save image for each FoV
         if multiple_fov:
-            output_name_originalimage = os.path.join(output_path, image.name+"_FoV"+str(f+1)+".tif")
+            output_name_originalimage = os.path.join(output_path, output_basename+"_FoV"+str(f+1)+".tif")
             fov_image = image.get_TYXarray()
-            fov_image = fov_image[:, np.newaxis, : ,:]
+            fov_image = fov_image[:, np.newaxis, :, :]
             tifffile.imwrite(output_name_originalimage, fov_image, metadata={'axes': 'TCYX'}, imagej=True, compression='zlib')
-            output_name = os.path.join(output_path, image.name+"_FoV"+str(f+1)+"_mask.tif")
+            output_name = os.path.join(output_path, output_basename+"_FoV"+str(f+1)+"_mask.tif")
         else:
-            output_name = os.path.join(output_path, image.name+"_mask.tif")
+            output_name = os.path.join(output_path, output_basename+"_mask.tif")
         # Save the mask
-        mask = mask[:, np.newaxis, : ,:]
+        mask = mask[:, np.newaxis, :, :]
         tifffile.imwrite(output_name, mask, metadata={'axes': 'TCYX'}, imagej=True, compression='zlib')
 
         logger.info("Saving segmentation masks to %s", output_name)
@@ -123,14 +125,14 @@ def main(image_path, model_path, output_path, display_results=True, use_gpu=True
         if display_results:
             QMessageBox.information(viewer_images.window._qt_window, 'File saved', 'Masks saved to\n' + output_name)
 
-    if display_results:  
-            # Stop logging into napari window & restore cursor
-            napari.qt.get_app().restoreOverrideCursor()
-            viewer_images.window._status_bar._toggle_activity_dock(False)
-            pbr.close()
-            # Show mask in napari
-            layer_mask = viewer_images.add_labels(mask, name="Cell mask")
-            layer_mask.editable = False # Do not allow edition
+    if display_results:
+        # Stop logging into napari window & restore cursor
+        napari.qt.get_app().restoreOverrideCursor()
+        viewer_images.window._status_bar._toggle_activity_dock(False)
+        pbr.close()
+        # Show mask in napari
+        layer_mask = viewer_images.add_labels(mask, name="Cell mask")
+        layer_mask.editable = False
 
     # stop using logfile
     logger.removeHandler(logfile_handler)