+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 * Z-projection module: check that output filenames are unique (i.e. to avoid two input files with same output filename).
+* Segmentation module: check that output filenames are unique (i.e. to avoid two input files with same output filename).
 
 ### Changed